+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Missing check of `lt_handle_t.l3.session_status` in `lt_in__ecc_key_generate()`.
 - CAL: support for OpenSSL.
 - CAL: support for WolfCrypt.
+- Linux HAL: added new HAL for Linux which utilizes spidev for chip select instead of GPIO.
 
 ### Fixed
 - `lt_print_bytes` function now returns `LT_PARAM_ERR` when incorrect parameters are passed instead of `LT_FAIL`.